Refactor diagram widget for better readability
