+++ src/view/widgets/DiagramWidget/diagram/DiagramWidget.jsx
@@ -1,43 +1,15 @@
-import React, { useCallback, useEffect, useState, useRef } from 'react';
+import React, { useRef } from 'react';
 import { CanvasWidget } from '@projectstorm/react-canvas-core';
-import { NodeModel } from '@projectstorm/react-diagrams';
-
-import { useDrop } from 'react-dnd';
 import Box from '@mui/material/Box';
-import debounce from 'lodash/debounce';
-import PropTypes from 'prop-types';
-import createRadicalEngine, {
-  updateRadicalEngine,
-} from './core/createRadicalEngine';
 
-import {
-  DRAG_DIAGRAM_ITEMS_END_EVENT,
-  LINK_CONNECTED_TO_TARGET_EVENT,
-  DIAGRAM_ENTITY_SELECTED,
-  DIAGRAM_NODE_COLLAPSED,
-  DIAGRAM_NODE_EXPANDED,
-  DIAGRAM_ENTITY_DELETED,
-  MODEL_DROP_TYPE,
-  METAMODEL_DROP_TYPE,
-  DIAGRAM_LINK_TARGET_SELECTED_EVENT,
-  DIAGRAM_ITEM_NAME_CHANGED,
-  DIAGRAM_NODE_DETACHED,
-  DRAG_CANVAS_END_EVENT,
-  CANVAS_ZOOM_CHANGED,
-} from './consts';
-import { addLinks, addNodes } from './core/viewModelRenderer';
-import RadicalDiagramModel from './core/RadicalDiagramModel';
+import PropTypes from 'prop-types';
+import { CANVAS_ZOOM_CHANGED } from './consts';
 import ToolbarMenu from '../../../components/ToolbarMenu';
 import { getCanvas } from '../../../../tests/getDataTestId';
 import ObjectMenuWrapper from './components/ObjectMenuWrapper';
-
-const zoomDebounceTime = 500;
-const mapViewmodel = (viewmodel, editMode) => {
-  const diagramModel = new RadicalDiagramModel();
-  addNodes(diagramModel, viewmodel, editMode);
-  addLinks(diagramModel, viewmodel, editMode);
-  return diagramModel;
-};
+import useCallbackRegistration from './hooks/useCallbackRegistration';
+import useViewSetup from './hooks/useViewSetup';
+import useDragAndDrop from './hooks/useDragAndDrop';
 
 const fillStyle = {
   width: '100%',
@@ -78,154 +50,34 @@ const DiagramWidget = ({
   setLinkingMode,
 }) => {
   const contextMenuCallbackRef = useRef(null);
-  const debouncedZoom = debounce(onDiagramAlignmentUpdated, zoomDebounceTime);
-  const registerCallbacks = useCallback(
-    () => ({
-      eventDidFire: (e) => {
-        switch (e.function) {
-          case DRAG_DIAGRAM_ITEMS_END_EVENT:
-            onDragItemsEnd({ ...e.point }, e.items);
-            break;
-          case LINK_CONNECTED_TO_TARGET_EVENT:
-            onLinkConnected(e.sourceId, e.targetId);
-            break;
-          case DRAG_CANVAS_END_EVENT:
-            onDiagramAlignmentUpdated(e.offsetX, e.offsetY, e.zoom);
-            break;
-          case CANVAS_ZOOM_CHANGED:
-            debouncedZoom(e.offsetX, e.offsetY, e.zoom);
-            break;
-          case DIAGRAM_ENTITY_SELECTED:
-            onItemSelected(
-              e.entity.getID(),
-              e.entity instanceof NodeModel ? 'object' : 'relation',
-              e.isSelected
-            );
-            break;
-          case DIAGRAM_NODE_COLLAPSED:
-            onNodeCollapsed(e.id);
-            break;
-          case DIAGRAM_NODE_EXPANDED:
-            onNodeExpanded(e.id);
-            break;
-          case DIAGRAM_ENTITY_DELETED:
-            if (e.entity instanceof NodeModel) {
-              if (e.deleteFromModel) {
-                onObjectRemove(e.entity.getID());
-              } else {
-                onNodeRemove(e.entity.getID());
-              }
-            } else if (e.deleteFromModel) {
-              onRelationRemove(e.entity.getID());
-            } else {
-              onLinkRemove(e.entity.getID());
-            }
-            break;
-          case DIAGRAM_LINK_TARGET_SELECTED_EVENT:
-            onLinkConnected(e.id, e.source, e.target, e.type);
-            break;
-          case DIAGRAM_ITEM_NAME_CHANGED:
-            onItemNameUpdated(
-              e.entity.getID(),
-              e.entity instanceof NodeModel ? 'object' : 'relation',
-              e.entity.options.name
-            );
-            break;
-          case DIAGRAM_NODE_DETACHED:
-            onNodeDetached(e.id);
-            break;
-          default:
-            break;
-        }
-      },
-    }),
-    [
-      onDragItemsEnd,
-      onLinkConnected,
-      onDiagramAlignmentUpdated,
-      onItemSelected,
-      onNodeExpanded,
-      onNodeCollapsed,
-      onNodeRemove,
-      onLinkRemove,
-      onObjectRemove,
-      onRelationRemove,
-      onItemNameUpdated,
-      onNodeDetached,
-      debouncedZoom,
-    ]
+  const registerCallbacks = useCallbackRegistration(
+    onDragItemsEnd,
+    onLinkConnected,
+    onDiagramAlignmentUpdated,
+    onItemSelected,
+    onNodeCollapsed,
+    onNodeExpanded,
+    onObjectRemove,
+    onNodeRemove,
+    onRelationRemove,
+    onLinkRemove,
+    onItemNameUpdated,
+    onNodeDetached
   );
-
-  const [engine] = useState(createRadicalEngine());
-  const [isModelSet, setIsModelSet] = useState(false);
-  const [viewName, setViewName] = useState();
-  useEffect(() => {
-    updateRadicalEngine(engine, editEnabled, selectionEnabled);
-    const isViewChanged = viewName !== viewmodel.name;
-    setViewName(viewmodel.name);
-    const model = mapViewmodel(viewmodel, editEnabled);
-
-    model.registerListener(registerCallbacks());
-    model.getNodes().forEach((node) => {
-      node.registerListener(registerCallbacks());
-    });
-    model.getLinks().forEach((link) => {
-      link.registerListener(registerCallbacks());
-      link.update();
-    });
-    if (smoothTransitionMode && !isViewChanged) {
-      const sourceZoomLevel = engine.getModel().getZoomLevel();
-      const sourceOffsetX = engine.getModel().getOffsetX();
-      const sourceOffsetY = engine.getModel().getOffsetY();
-      engine.setModel(model);
-      engine.moveWithAnim(
-        sourceZoomLevel,
-        sourceOffsetX,
-        sourceOffsetY,
-        alignment.zoom,
-        alignment.offsetX,
-        alignment.offsetY
-      );
-    } else {
-      engine.setModel(model);
-      model.setInitialZoomLevel(alignment.zoom);
-      model.setInitialOffset(alignment.offsetX, alignment.offsetY);
-    }
-    setIsModelSet(true);
-  }, [
+  const [engine, isModelSet] = useViewSetup(
+    registerCallbacks,
     viewmodel,
     editEnabled,
-    smoothTransitionMode,
-    alignment,
-    engine,
-    registerCallbacks,
-    viewName,
     selectionEnabled,
-  ]);
+    smoothTransitionMode,
+    alignment
+  );
 
-  const [, drop] = useDrop(() => ({
-    accept: [MODEL_DROP_TYPE, METAMODEL_DROP_TYPE],
-    drop: (item, monitor) => {
-      const dropPoint = engine.getRelativeMousePoint({
-        clientX: monitor.getClientOffset().x,
-        clientY: monitor.getClientOffset().y,
-      });
-
-      const node = engine.getNodeAtPosition(
-        monitor.getClientOffset().x,
-        monitor.getClientOffset().y
-      );
-      if (item.type === MODEL_DROP_TYPE) {
-        onAddObjectToView(item.id, { ...dropPoint });
-      } else {
-        onAddMetamodelObjectToView(
-          item.metamodelType,
-          { ...dropPoint },
-          node ? node.getID() : undefined
-        );
-      }
-    },
-  }));
+  const [, drop] = useDragAndDrop(
+    engine,
+    onAddObjectToView,
+    onAddMetamodelObjectToView
+  );
   return (
     engine &&
     isModelSet && (
